refactor: introduce market domain for price-change display
