+++ src/features/discover/components/markets/cards/MarketPill.tsx
@@ -13,8 +13,8 @@ import { opacity } from '@/design-system/utils/opacity';
 import { MarketIcon } from '@/features/discover/components/markets/cards/MarketIcon';
 import { MarketPriceChange } from '@/features/discover/components/markets/cards/MarketPriceChange';
 import { type MarketDisplayItem } from '@/features/discover/types/marketDisplayItem';
-import { parseNormalizedPercentChange } from '@/framework/ui/price/formatPriceChange';
-import { usePriceChangeColors } from '@/framework/ui/price/usePriceChangeColors';
+import { usePriceChangeColors } from '@/features/market/ui/hooks/usePriceChangeColors';
+import { parseNormalizedPercentChange } from '@/features/market/ui/utils/formatPriceChange';
 import { THICKER_BORDER_WIDTH } from '@/styles/constants';
 import { measureTextSync, type MeasureTextProps } from '@/utils/measureText';
 import { getHighContrastTextColorWorklet } from '@/worklets/colors';
