+++ src/features/discover/components/markets/cards/PredictionMarketTileCard.tsx
@@ -11,12 +11,13 @@ import { LiveTokenText } from '@/components/live-token-text/LiveTokenText';
 import { globalColors, Text, useColorMode } from '@/design-system';
 import { opacity } from '@/design-system/utils/opacity';
 import { type CardPressHandler, type OrderPressHandler } from '@/features/discover/types/sectionLayout';
+import { usePriceChangeColors } from '@/features/market/ui/hooks/usePriceChangeColors';
+import { getPriceChangeColor } from '@/features/market/ui/utils/priceChangeColors';
 import { DOWN_ARROW, UP_ARROW } from '@/features/perps/constants';
 import { type PolymarketEvent, type PolymarketMarket } from '@/features/polymarket/types/polymarket-event';
 import { getOutcomeColor } from '@/features/polymarket/utils/getMarketColor';
 import { formatOdds } from '@/features/polymarket/utils/sportsEventBetData';
 import { toPercentageWorklet } from '@/framework/core/safeMath';
-import { getPriceChangeColor, usePriceChangeColors } from '@/framework/ui/price/usePriceChangeColors';
 import { formatNumber } from '@/helpers/strings';
 import Navigation from '@/navigation/Navigation';
 import Routes from '@/navigation/routesNames';
