+++ src/features/discover/components/markets/hooks/usePerpMarketDisplay.ts
@@ -4,13 +4,13 @@ import { globalColors } from '@/design-system/color/palettes';
 import { opacity } from '@/design-system/utils/opacity';
 import { type MarketPillWidthInput } from '@/features/discover/components/markets/cards/MarketPill';
 import { type MarketDisplayItem } from '@/features/discover/types/marketDisplayItem';
+import { getPriceChangeColor, getPriceChangeColors } from '@/features/market/ui/utils/priceChangeColors';
 import { HYPERLIQUID_COLORS } from '@/features/perps/constants';
 import { useHyperliquidLineChartsStore } from '@/features/perps/stores/hyperliquidLineChartsStore';
 import { type PerpMarketWithMetadata } from '@/features/perps/types';
 import { convertStoredPerpPriceChangeToPercent, getHyperliquidTokenId } from '@/features/perps/utils';
 import { formatPerpAssetPrice } from '@/features/perps/utils/formatPerpsAssetPrice';
 import { type PerpMarketPlacementItem } from '@/features/placements/stores/derived/perpsPlacementStore';
-import { getPriceChangeColor, getPriceChangeColors } from '@/framework/ui/price/usePriceChangeColors';
 
 const MARKET_NEUTRAL_CHART_COLOR = opacity(globalColors.white100, 0.5);
 const MARKET_CHART_PRICE_CHANGE_COLORS = {
