+++ src/features/market/ui/utils/formatPriceChange.test.ts
